moved supportcontact to support app
