+++ app/apps/support/admin.py
@@ -1,0 +1,5 @@
+from django.contrib import admin
+
+from .models import SupportContact
+
+admin.site.register(SupportContact, admin.ModelAdmin)
