+++ app/apps/support/apps.py
@@ -1,0 +1,5 @@
+from django.apps import AppConfig
+
+
+class SupportConfig(AppConfig):
+    name = "support"
